Remove baseWebviewElement
This is no longer needed now what we only have iframe base webviews

This change merges `IframeWebviewElement` into `BaseWebviewElement`
